# Simplify the histogram aggregator