--- a/sdk/metric/internal/aggregate/histogram_test.go
+++ b/sdk/metric/internal/aggregate/histogram_test.go
@@ -15,6 +15,7 @@
 package aggregate // import "go.opentelemetry.io/otel/sdk/metric/internal/aggregate"
 
 import (
+	"context"
 	"sort"
 	"testing"
 
@@ -37,74 +38,155 @@ var (
 
 func TestHistogram(t *testing.T) {
 	t.Cleanup(mockTime(now))
-	t.Run("Int64", testHistogram[int64])
-	t.Run("Float64", testHistogram[float64])
+
+	t.Run("Int64/Delta/Sum", testDeltaHist[int64](conf[int64]{hPt: hPointSummed[int64]}))
+	t.Run("Int64/Delta/NoSum", testDeltaHist[int64](conf[int64]{noSum: true, hPt: hPoint[int64]}))
+	t.Run("Float64/Delta/Sum", testDeltaHist[float64](conf[float64]{hPt: hPointSummed[float64]}))
+	t.Run("Float64/Delta/NoSum", testDeltaHist[float64](conf[float64]{noSum: true, hPt: hPoint[float64]}))
+
+	t.Run("Int64/Cumulative/Sum", testCumulativeHist[int64](conf[int64]{hPt: hPointSummed[int64]}))
+	t.Run("Int64/Cumulative/NoSum", testCumulativeHist[int64](conf[int64]{noSum: true, hPt: hPoint[int64]}))
+	t.Run("Float64/Cumulative/Sum", testCumulativeHist[float64](conf[float64]{hPt: hPointSummed[float64]}))
+	t.Run("Float64/Cumulative/NoSum", testCumulativeHist[float64](conf[float64]{noSum: true, hPt: hPoint[float64]}))
 }
 
-func testHistogram[N int64 | float64](t *testing.T) {
-	tester := &aggregatorTester[N]{
-		GoroutineN:   defaultGoroutines,
-		MeasurementN: defaultMeasurements,
-		CycleN:       defaultCycles,
-	}
-
-	incr := monoIncr[N]()
-	eFunc := deltaHistSummedExpecter[N](incr)
-	t.Run("Delta/Summed", tester.Run(newDeltaHistogram[N](histConf, false), incr, eFunc))
-	eFunc = deltaHistExpecter[N](incr)
-	t.Run("Delta/NoSum", tester.Run(newDeltaHistogram[N](histConf, true), incr, eFunc))
-	eFunc = cumuHistSummedExpecter[N](incr)
-	t.Run("Cumulative/Summed", tester.Run(newCumulativeHistogram[N](histConf, false), incr, eFunc))
-	eFunc = cumuHistExpecter[N](incr)
-	t.Run("Cumulative/NoSum", tester.Run(newCumulativeHistogram[N](histConf, true), incr, eFunc))
+type conf[N int64 | float64] struct {
+	noSum bool
+	hPt   func(attribute.Set, N, uint64) metricdata.HistogramDataPoint[N]
 }
 
-func deltaHistSummedExpecter[N int64 | float64](incr setMap[N]) expectFunc {
-	h := metricdata.Histogram[N]{Temporality: metricdata.DeltaTemporality}
-	return func(m int) metricdata.Aggregation {
-		h.DataPoints = make([]metricdata.HistogramDataPoint[N], 0, len(incr))
-		for a, v := range incr {
-			h.DataPoints = append(h.DataPoints, hPointSummed[N](a, v, uint64(m)))
-		}
-		return h
-	}
+func testDeltaHist[N int64 | float64](c conf[N]) func(t *testing.T) {
+	in, out := Builder[N]{
+		Temporality: metricdata.DeltaTemporality,
+		Filter:      attrFltr,
+	}.ExplicitBucketHistogram(histConf, c.noSum)
+	ctx := context.Background()
+	return test[N](in, out, []teststep[N]{
+		{
+			input: []arg[N]{},
+			expect: output{
+				n: 0,
+				agg: metricdata.Histogram[N]{
+					Temporality: metricdata.DeltaTemporality,
+					DataPoints:  []metricdata.HistogramDataPoint[N]{},
+				},
+			},
+		},
+		{
+			input: []arg[N]{
+				{ctx, 2, alice},
+				{ctx, 10, bob},
+				{ctx, 2, alice},
+				{ctx, 2, alice},
+				{ctx, 10, bob},
+			},
+			expect: output{
+				n: 2,
+				agg: metricdata.Histogram[N]{
+					Temporality: metricdata.DeltaTemporality,
+					DataPoints: []metricdata.HistogramDataPoint[N]{
+						c.hPt(fltrAlice, 2, 3),
+						c.hPt(fltrBob, 10, 2),
+					},
+				},
+			},
+		},
+		{
+			input: []arg[N]{
+				{ctx, 10, alice},
+				{ctx, 3, bob},
+			},
+			expect: output{
+				n: 2,
+				agg: metricdata.Histogram[N]{
+					Temporality: metricdata.DeltaTemporality,
+					DataPoints: []metricdata.HistogramDataPoint[N]{
+						c.hPt(fltrAlice, 10, 1),
+						c.hPt(fltrBob, 3, 1),
+					},
+				},
+			},
+		},
+		{
+			input: []arg[N]{},
+			// Delta histograms are expected to reset.
+			expect: output{
+				n: 0,
+				agg: metricdata.Histogram[N]{
+					Temporality: metricdata.DeltaTemporality,
+					DataPoints:  []metricdata.HistogramDataPoint[N]{},
+				},
+			},
+		},
+	})
 }
 
-func deltaHistExpecter[N int64 | float64](incr setMap[N]) expectFunc {
-	h := metricdata.Histogram[N]{Temporality: metricdata.DeltaTemporality}
-	return func(m int) metricdata.Aggregation {
-		h.DataPoints = make([]metricdata.HistogramDataPoint[N], 0, len(incr))
-		for a, v := range incr {
-			h.DataPoints = append(h.DataPoints, hPoint[N](a, v, uint64(m)))
-		}
-		return h
-	}
-}
-
-func cumuHistSummedExpecter[N int64 | float64](incr setMap[N]) expectFunc {
-	var cycle int
-	h := metricdata.Histogram[N]{Temporality: metricdata.CumulativeTemporality}
-	return func(m int) metricdata.Aggregation {
-		cycle++
-		h.DataPoints = make([]metricdata.HistogramDataPoint[N], 0, len(incr))
-		for a, v := range incr {
-			h.DataPoints = append(h.DataPoints, hPointSummed[N](a, v, uint64(cycle*m)))
-		}
-		return h
-	}
-}
-
-func cumuHistExpecter[N int64 | float64](incr setMap[N]) expectFunc {
-	var cycle int
-	h := metricdata.Histogram[N]{Temporality: metricdata.CumulativeTemporality}
-	return func(m int) metricdata.Aggregation {
-		cycle++
-		h.DataPoints = make([]metricdata.HistogramDataPoint[N], 0, len(incr))
-		for a, v := range incr {
-			h.DataPoints = append(h.DataPoints, hPoint[N](a, v, uint64(cycle*m)))
-		}
-		return h
-	}
+func testCumulativeHist[N int64 | float64](c conf[N]) func(t *testing.T) {
+	in, out := Builder[N]{
+		Temporality: metricdata.CumulativeTemporality,
+		Filter:      attrFltr,
+	}.ExplicitBucketHistogram(histConf, c.noSum)
+	ctx := context.Background()
+	return test[N](in, out, []teststep[N]{
+		{
+			input: []arg[N]{},
+			expect: output{
+				n: 0,
+				agg: metricdata.Histogram[N]{
+					Temporality: metricdata.CumulativeTemporality,
+					DataPoints:  []metricdata.HistogramDataPoint[N]{},
+				},
+			},
+		},
+		{
+			input: []arg[N]{
+				{ctx, 2, alice},
+				{ctx, 10, bob},
+				{ctx, 2, alice},
+				{ctx, 2, alice},
+				{ctx, 10, bob},
+			},
+			expect: output{
+				n: 2,
+				agg: metricdata.Histogram[N]{
+					Temporality: metricdata.CumulativeTemporality,
+					DataPoints: []metricdata.HistogramDataPoint[N]{
+						c.hPt(fltrAlice, 2, 3),
+						c.hPt(fltrBob, 10, 2),
+					},
+				},
+			},
+		},
+		{
+			input: []arg[N]{
+				{ctx, 2, alice},
+				{ctx, 10, bob},
+			},
+			expect: output{
+				n: 2,
+				agg: metricdata.Histogram[N]{
+					Temporality: metricdata.CumulativeTemporality,
+					DataPoints: []metricdata.HistogramDataPoint[N]{
+						c.hPt(fltrAlice, 2, 4),
+						c.hPt(fltrBob, 10, 3),
+					},
+				},
+			},
+		},
+		{
+			input: []arg[N]{},
+			expect: output{
+				n: 2,
+				agg: metricdata.Histogram[N]{
+					Temporality: metricdata.CumulativeTemporality,
+					DataPoints: []metricdata.HistogramDataPoint[N]{
+						c.hPt(fltrAlice, 2, 4),
+						c.hPt(fltrBob, 10, 3),
+					},
+				},
+			},
+		},
+	})
 }
 
 // hPointSummed returns an HistogramDataPoint that started and ended now with
@@ -190,93 +272,89 @@ func testBucketsSum[N int64 | float64]() func(t *testing.T) {
 	}
 }
 
-func testHistImmutableBounds[N int64 | float64](newA func(aggregation.ExplicitBucketHistogram, bool) aggregator[N], getBounds func(aggregator[N]) []float64) func(t *testing.T) {
+func TestHistogramImmutableBounds(t *testing.T) {
 	b := []float64{0, 1, 2}
 	cpB := make([]float64, len(b))
 	copy(cpB, b)
 
-	a := newA(aggregation.ExplicitBucketHistogram{Boundaries: b}, false)
-	return func(t *testing.T) {
-		require.Equal(t, cpB, getBounds(a))
+	h := newHistogram[int64](aggregation.ExplicitBucketHistogram{Boundaries: b}, false)
+	require.Equal(t, cpB, h.bounds)
 
-		b[0] = 10
-		assert.Equal(t, cpB, getBounds(a), "modifying the bounds argument should not change the bounds")
+	b[0] = 10
+	assert.Equal(t, cpB, h.bounds, "modifying the bounds argument should not change the bounds")
 
-		a.Aggregate(5, alice)
-		hdp := a.Aggregation().(metricdata.Histogram[N]).DataPoints[0]
-		hdp.Bounds[1] = 10
-		assert.Equal(t, cpB, getBounds(a), "modifying the Aggregation bounds should not change the bounds")
-	}
-}
+	h.measure(context.Background(), 5, alice)
 
-func TestHistogramImmutableBounds(t *testing.T) {
-	t.Run("Delta", testHistImmutableBounds(
-		newDeltaHistogram[int64],
-		func(a aggregator[int64]) []float64 {
-			deltaH := a.(*deltaHistogram[int64])
-			return deltaH.bounds
-		},
-	))
-
-	t.Run("Cumulative", testHistImmutableBounds(
-		newCumulativeHistogram[int64],
-		func(a aggregator[int64]) []float64 {
-			cumuH := a.(*cumulativeHistogram[int64])
-			return cumuH.bounds
-		},
-	))
+	var data metricdata.Aggregation = metricdata.Histogram[int64]{}
+	h.cumulative(&data)
+	hdp := data.(metricdata.Histogram[int64]).DataPoints[0]
+	hdp.Bounds[1] = 10
+	assert.Equal(t, cpB, h.bounds, "modifying the Aggregation bounds should not change the bounds")
 }
 
 func TestCumulativeHistogramImutableCounts(t *testing.T) {
-	a := newCumulativeHistogram[int64](histConf, false)
-	a.Aggregate(5, alice)
-	hdp := a.Aggregation().(metricdata.Histogram[int64]).DataPoints[0]
+	h := newHistogram[int64](histConf, false)
+	h.measure(context.Background(), 5, alice)
 
-	cumuH := a.(*cumulativeHistogram[int64])
-	require.Equal(t, hdp.BucketCounts, cumuH.values[alice].counts)
+	var data metricdata.Aggregation = metricdata.Histogram[int64]{}
+	h.cumulative(&data)
+	hdp := data.(metricdata.Histogram[int64]).DataPoints[0]
+
+	require.Equal(t, hdp.BucketCounts, h.values[alice].counts)
 
 	cpCounts := make([]uint64, len(hdp.BucketCounts))
 	copy(cpCounts, hdp.BucketCounts)
 	hdp.BucketCounts[0] = 10
-	assert.Equal(t, cpCounts, cumuH.values[alice].counts, "modifying the Aggregator bucket counts should not change the Aggregator")
+	assert.Equal(t, cpCounts, h.values[alice].counts, "modifying the Aggregator bucket counts should not change the Aggregator")
 }
 
 func TestDeltaHistogramReset(t *testing.T) {
 	t.Cleanup(mockTime(now))
 
-	a := newDeltaHistogram[int64](histConf, false)
-	assert.Nil(t, a.Aggregation())
+	h := newHistogram[int64](histConf, false)
 
-	a.Aggregate(1, alice)
+	var data metricdata.Aggregation = metricdata.Histogram[int64]{}
+	require.Equal(t, 0, h.delta(&data))
+	require.Len(t, data.(metricdata.Histogram[int64]).DataPoints, 0)
+
+	h.measure(context.Background(), 1, alice)
+
 	expect := metricdata.Histogram[int64]{Temporality: metricdata.DeltaTemporality}
 	expect.DataPoints = []metricdata.HistogramDataPoint[int64]{hPointSummed[int64](alice, 1, 1)}
-	metricdatatest.AssertAggregationsEqual(t, expect, a.Aggregation())
+	h.delta(&data)
+	metricdatatest.AssertAggregationsEqual(t, expect, data)
 
 	// The attr set should be forgotten once Aggregations is called.
 	expect.DataPoints = nil
-	assert.Nil(t, a.Aggregation())
+	assert.Equal(t, 0, h.delta(&data))
+	assert.Len(t, data.(metricdata.Histogram[int64]).DataPoints, 0)
 
 	// Aggregating another set should not affect the original (alice).
-	a.Aggregate(1, bob)
+	h.measure(context.Background(), 1, bob)
 	expect.DataPoints = []metricdata.HistogramDataPoint[int64]{hPointSummed[int64](bob, 1, 1)}
-	metricdatatest.AssertAggregationsEqual(t, expect, a.Aggregation())
-}
-
-func TestEmptyHistogramNilAggregation(t *testing.T) {
-	assert.Nil(t, newCumulativeHistogram[int64](histConf, false).Aggregation())
-	assert.Nil(t, newCumulativeHistogram[float64](histConf, false).Aggregation())
-	assert.Nil(t, newDeltaHistogram[int64](histConf, false).Aggregation())
-	assert.Nil(t, newDeltaHistogram[float64](histConf, false).Aggregation())
+	h.delta(&data)
+	metricdatatest.AssertAggregationsEqual(t, expect, data)
 }
 
 func BenchmarkHistogram(b *testing.B) {
-	b.Run("Int64", benchmarkHistogram[int64])
-	b.Run("Float64", benchmarkHistogram[float64])
-}
-
-func benchmarkHistogram[N int64 | float64](b *testing.B) {
-	factory := func() aggregator[N] { return newDeltaHistogram[N](histConf, false) }
-	b.Run("Delta", benchmarkAggregator(factory))
-	factory = func() aggregator[N] { return newCumulativeHistogram[N](histConf, false) }
-	b.Run("Cumulative", benchmarkAggregator(factory))
+	b.Run("Int64/Cumulative", benchmarkAggregate(func() (Measure[int64], ComputeAggregation) {
+		return Builder[int64]{
+			Temporality: metricdata.CumulativeTemporality,
+		}.ExplicitBucketHistogram(histConf, false)
+	}))
+	b.Run("Int64/Delta", benchmarkAggregate(func() (Measure[int64], ComputeAggregation) {
+		return Builder[int64]{
+			Temporality: metricdata.DeltaTemporality,
+		}.ExplicitBucketHistogram(histConf, false)
+	}))
+	b.Run("Float64/Cumulative", benchmarkAggregate(func() (Measure[float64], ComputeAggregation) {
+		return Builder[float64]{
+			Temporality: metricdata.CumulativeTemporality,
+		}.ExplicitBucketHistogram(histConf, false)
+	}))
+	b.Run("Float64/Delta", benchmarkAggregate(func() (Measure[float64], ComputeAggregation) {
+		return Builder[float64]{
+			Temporality: metricdata.DeltaTemporality,
+		}.ExplicitBucketHistogram(histConf, false)
+	}))
 }