… sdk/metric/internal/aggregate/aggregator.go
… sdk/metric/internal/aggregate/aggregator_test.go